fix: address PR #22 review round 2 -- 5 issues
- messages.rs: use From<u64>/From<u32>/From<u8> for ciborium Integer
  (no more i64 cast, full u64 range preserved)
- messages.rs: COMMAND payload nesting -- command-specific fields are
  now inside a nested CBOR map under key 5 (payload), matching the
  protocol spec. NOP/REBOOT omit key 5 entirely.
- chunk.rs: chunk_count uses u32::try_from instead of as u32 (returns
  None on overflow)
- Cargo.toml: remove misleading alloc feature (crate unconditionally
  requires alloc); keep std feature for future use
- program_image.rs: use From<u64>/From<u32> for Integer encoding
- protocol.md: clarify that COMMAND payload (key 5) is a nested CBOR
  map containing command-specific fields, omitted for NOP/REBOOT

+++ crates/sonde-protocol/src/program_image.rs
@@ -9,7 +9,6 @@ use crate::constants::*;
 use crate::error::{DecodeError, EncodeError};
 use crate::traits::Sha256Provider;
 
-#[cfg(feature = "alloc")]
 use alloc::format;
 
 #[derive(Debug, Clone, PartialEq)]
@@ -37,20 +36,20 @@ impl ProgramImage {
                 // Keys in ascending order: 1, 2, 3, 4
                 Value::Map(alloc::vec![
                     (
-                        Value::Integer((MAP_KEY_TYPE as i64).into()),
-                        Value::Integer((m.map_type as i64).into()),
+                        Value::Integer(MAP_KEY_TYPE.into()),
+                        Value::Integer(m.map_type.into()),
                     ),
                     (
-                        Value::Integer((MAP_KEY_KEY_SIZE as i64).into()),
-                        Value::Integer((m.key_size as i64).into()),
+                        Value::Integer(MAP_KEY_KEY_SIZE.into()),
+                        Value::Integer(m.key_size.into()),
                     ),
                     (
-                        Value::Integer((MAP_KEY_VALUE_SIZE as i64).into()),
-                        Value::Integer((m.value_size as i64).into()),
+                        Value::Integer(MAP_KEY_VALUE_SIZE.into()),
+                        Value::Integer(m.value_size.into()),
                     ),
                     (
-                        Value::Integer((MAP_KEY_MAX_ENTRIES as i64).into()),
-                        Value::Integer((m.max_entries as i64).into()),
+                        Value::Integer(MAP_KEY_MAX_ENTRIES.into()),
+                        Value::Integer(m.max_entries.into()),
                     ),
                 ])
             })
@@ -59,11 +58,11 @@ impl ProgramImage {
         // Outer map keys in ascending order: 1 (bytecode), 2 (maps)
         let outer = Value::Map(alloc::vec![
             (
-                Value::Integer((IMG_KEY_BYTECODE as i64).into()),
+                Value::Integer(IMG_KEY_BYTECODE.into()),
                 Value::Bytes(self.bytecode.clone()),
             ),
             (
-                Value::Integer((IMG_KEY_MAPS as i64).into()),
+                Value::Integer(IMG_KEY_MAPS.into()),
                 Value::Array(map_values),
             ),
         ]);
